Improve transitions and face tracking stability
More stable face tracking was implemented in the detection center.
The utilization of a distance from the mean was used and about 5-10
areas are kept and continually replaced by new areas that are accepted.
A transition was implemented after a table view cell is selected.  The
transition delegate was used. Animations to views were implemented for a
more interactive transition.  Improvements to generics in queue and
protocols were improved upon.